This closes #229
GEODE-11: Fixing bug in definedIndexMap in LuceneServiceImpl
If an index is created on an existing region, an exception is thrown, but the index is still stored in definedIndexMap. This is fixed by removing the index definition before the exception is thrown. Added a test to verify

GEODE-11: Describe/Search index gfsh commands returns error message when region not created

- Previously, the describe/search index gfsh command on definied indexes returned an exception. Changed the behavior to return a message. Added dunit test to verify
- Changed Search index command to "search lucene" (previously "lucene search")