GEODE 11: Added a check to make sure StandardAnalyzer and NULL were considered the same.
LuceneIndexCreationProfile checks to makes sure the newly created index's field analyzers are consistent with existing indexes' field analyzers. However, when a field's analyzer is null and a new index is created with StandardAnalyzer on the same field, it failed. Fixed this issue by modifying the conditional checks. Added dunit tests to verify.

Signed-off-by: Gester Zhou <gzhou@pivotal.io>