Fixed eclipse compilation and test execution issues
With the eclipse compiler, AsyncEventListenerDUnitTest was generating a
compilation error about a deserializeLambda method being to large. This
doesn't seem to be a real issue because the javac doesn't have
the same problem. However, by refactoring the test slightly I was able
to make this error go away.

Changed the eclipse classpath to include the generated-resources
directories. These directories contain files like
GemFireVersion.properties that need to be included on the classpath, but
we were filtering them out.