GEODE-1836: remove defunct Java 1.6 block for Mac
Geode requires Java 1.8 and does not support Java 1.6. Attempting to
add the 1.6 classes.jar to the classpath for the Attach API results in
breaking the start and status commands in GFSH.

Removal of these script lines allows GFSH to use either the Attach API
in Java 1.8 or an alternative implementation for start and status. The
alternative implementation uses temporary files to communcate with the
other process. If you want to make sure you are using the Attach API
then simply set JAVA_HOME to the location of JDK 1.8.

If you use jenv on Mac, then consider configuring jenv to export
JAVA_HOME by executing:

$ jenv enable-plugin export