GEODE-1885: Removed call to check readiness (region check) after the offheap region entry is released.
GEODE-1885: Missing subsctiption event with Offheap partitioned region during bucket rebalance.

During the trasaction commit on redundant bucket region, if the bucket region is moved, the call-back logic (to deliver subscription events) were not invoked due to check-readiness call with offheap region. The check-readiness throws exception, if the region is not found, which causes the code to return early without sending the subscription events.

In this scenario, calling check-readiness is not needed...