GEODE-1915: Prevent deadlock registering instantiators with gateways
Don't hold a lock while distributing instantiators. This prevents the
deadlock because incoming registrations won't wait for registrations
that are being distributed.

This change might cause instantiators to be distributed in a different
order that they were registered in, but that's ok because the order in
which different instantiators are registered is not important.