native-client-software-grant - Mark buckets as blacklisted (primaries and secondaries)
When server goes down we want to mark it's primary buckets as
blacklisted for bucket-wait-timeout [ms]. This is not enough, adding the
secondary buckets that are refered by this server should be added to
blacklist too.
Let's assume server1 contains primary buckets P{a} and secondary buckets
S{b}.
Server2 contains S{a} and P{b}.
Server1 goes down.
Transactions that go to server2 to P{b} will pass, but the redundant
S{b} will fail.
This creates an inconsistency of the data and to solve it we want to add
all the buckets of the server, primary and secondary, to the blacklist
when we fail to connect to a server.