GEODE-2011 Client code which clears pdx registry needs synchronization
We clear pdx registry in client when we re-connect to cluster. But
during that time other thread may end up using old registry while other
thread is clearing this. Thus we need to synchronize that. In current
code it happens through listener events and I modified that notification
to synchronize this.