GEODE-2026: fix flaky test
The test was initializing a value to a String,
then starting an async thread that would keep
changing it to an int,
it slept for 100ms and then assummed that the async
thread had changed the value to an Integer.
In rare cases the thread never was able to to its first
put so the value was still a String.

The test now waits for the async thread to change the value
to an Integer instead of doing a 100ms sleep.