GEODE-2064: Added check for system shutdown while handlling connect exception
While message send in progress, if the system gets shutdown (forced disconnect),
the send (message delivery to peers) reports connect exception and ignores
detecting/throwing SystemDisconnect exception.

This results in client missing an event and resulting in data mismatch between
client and server.

Made changes to throw "DistributedSystemDisconnectedException" if system is
shutting down. This will result in caller to retry the operation.