GEODE-2073 CI Failure: GMSHealthMonitorJUnitTest.testHMNextNeighborAfterTimeout
I modified the test to wait longer and not require that the "next neighbor"
be an exact member but merely a different one than the initial "next neigbor".

Also, the diagnostic test in the exception thrown by this method was
including the JoinLeave membership view but that is null in these
tests.  Instead it needs to show the HealthMonitor's membership view.
I added a method to GMSHealthMonitor to allow access to its view.