GEODE-1978: Waiting for the queue to drain in testReplicatedSerialProp...
Waiting to the queue to drain in this test. The test had an assertion
that the queue was not empty. But really, in the background the
dispatcher was draining the queue as fast as it good. The actually
expected behavior is that the queue will become empty if the remote side
destroys a region.