GEODE-2208 Document transactions limitation with mixed region types.
- First operation to a transactional system which has colocated
data and replicated/distributed regions must be to a
partitioned region (with colocated data), to set the transaction
host.  Note that this was already documented, but not in a
prominent place.  Now the basics of this info is repeated in
a less-buried location.
- Fixed poorly-working links within the subnav and fixed a
couple of markdown errors.  The errors were improperly
specified anchor tags.