Code cleanup
See https://reviews.apache.org/r/55017/

While investigating a couple of failures I ran across a few things
that needed to be fixed.

1. EndpointManagerImpl was skipping notification of listeners if there
was no member ID in an endpoint but the member ID is no longer used.
I removed the check.

2. During a Forced Disconnect CacheClientProxy was skipping a lot of
clean-up because it was expecting a CacheClosedException instead of
a CancelException. We should never catch CacheClosedException -
always catch CancelException.

3. When SSL is being used the P2P Reader Threads were not setting connection
attributes in the thread's name like we do with non-SSL reader threads.
This made debugging an SSL failure a bit more difficult.