GEODE-2253 Locator may fail to respond to a valid request
This introduces a retry loop in the locator when a handler for an
incoming message can't be found.  It waits for an amount of time
for the handler to be installed, which I've set to the locator-
wait-time or 5 seconds if that property hasn't been set.

I've also changed InternalLocator to always install the handler
for cluster configuration status so that you can query any locator
to see if it has a cluster configuration service and, if so, what
state it's in.