GEODE-2137 client membership IDs contain 17 bytes of useless information
Reverting the removal of UUIDs from fully serialized member IDs.  There is
administration code that caches serialized events containing these IDs.
The cached form is transmitted between nodes without regard to versioning
so it is impossible to track the chain of custody and know how to properly
deserialize a 1.1.0 ID with this change.  Version support needs to be
extended to CachedDeserializable classes if we're going to make this sort
of change, so that they know how to properly deserialize an object of
a versioned class.

Removal of UUIDs is still in effect for the compact form of membership ID
used in version tags and elsewhere since this information is not held in
serialized form in cache regions.