GEODE-2368 Need to fix log message in DirectChannel
Prior to the membership rewrite for v1.0 we used a different identifier
for the DirectChannel using the class "Stub".  In the rewrite we replaced
this with the InternalDistributedMember ID, but that isn't formed until
after the DirectChannel has been booted up and the MembershipManager has
finished joining the distributed system.

This change set modifies DirectChannel to use the SocketAddress of the
server socket it's created, which is equivalent to the old "Stub" identifier.