GEODE-2449. Don't throw exception on redis port of zero
when Redis service is not found.

We would like to have the ability to tell the difference between
settings that are unset and those that are set to zero, but for the
moment we can't. So that's how it is. It may be that we'll have to not
allow starting a Redis host by setting a port number of zero in config,
or that it will just be less reliable.