GEODE-2490: Avoid processing tombstone GC message in-line
Currently the tombstone message sent for replicas are getting
processed in-line instead of handing it to thread pool.
Based on the number of nodes in the cluster, this may take
long time to process, impacting other cache operation that
required to be processed in-line.

The change provided here enables tombstone messages to
be not processed in-line instead processed in separate
thread.