GEODE-2029 Trimming unused parts of Lucene from binary NOTICE
The lucene section of our NOTICE included some elements for jars that we
are not redistributing like some language specific analyzers and
stemmers. To reduce the burden on downstream projects, removing these
components from our NOTICE.

I did leave in the section on servlet-api, which appeared to come from
lucene, but in fact we do include servlet-api.jar in our binary
distribution.

Here are the specific notes on where stuff in the notice came from

icu4j,WordBreakTestUnicode - this is in  lucene-analyzers-icu.jar (we don't include)
junit - we're not distributing this jar
stempel - this is in lucene-analyzers-stempel (we don't include)
smartcn - this is in  lucene-analyzers-smartcn (we don't include)
kuromoji - this in in lucene-analyzers-kuromoji (we don't include)
Morfologik - this in in lucene-analyzers-morfologik (we don't include)