dd official Socket timeout parameter when connecting to servers/locators
When connecting from the client to the servers/locators, if the
servers/locators is not started, the connection can not be established
and a Socket timeout occurs.
This timeout value is 59 seconds by default.
This timeout value is too long. This timeout value can be done by
specifying the informal parameter "gemfire.PoolImpl.HANDSHAKE_TIMEOUT"
in java the system property, but we corresponded so that it can be
specified by official parameters.

Timeout values ââare determined and determined in the following
order of priority.

1. Java System Property:gemfire.PoolImpl.HANDSHAKE_TIMEOUT
2. Java System Property:gemfire.connect-timeout
3. gemfire.properties:connect-timeout
4. Default:59000 milli seconds