GEODE-2788: Add official Socket timeout parameter when connecting to servers/locators
When connecting from the client to the servers/locators, if the servers/locators is not started, the connection can not be established and a Socket timeout occurs.
This timeout value is 59 seconds by default. This timeout value is too long. This timeout value can be changed by specifying the unofficial parameter "gemfire.PoolImpl.HANDSHAKE_TIMEOUT" in java system property, but I corresponded so that it can be specified by official parameters.
Like the NativeClient, the official parameters should be specified by "connect-timeout" in gemfire.properties.

Timeout values are determined in the following order of priority.

1. java system property:gemfire.PoolImpl.HANDSHAKE_TIMEOUT
2. java system property:gemfire.connect-timeout
3. gemfire.properties:connect-timeout
4. default:59000 milli seconds