GEODE-2632: fix ParallelQueueRemovalMessageJUnitTest
Fakes.cache() mocks GemFireCacheImpl without specifying a "when"
for getRegion(String). When that method is final, Mockito can't
override it, so it ends up calling getRegion(String, boolean)
which is overridden by the test. When it's non-final Mockito
overrides it but the test doesn't specify a "when" so the
getRegion(String) method becomes a no-op.

This change restores final on getRegion(String) and removes
@Ignore from the test.