GEODE-2802 Tombstone version vector to contain only the members that generate the tombstone
TombstoneMessage serialization code assumes the member info in RVV to be either
membership-id or disk-id and uses this info while de-serializing.
When there is a mix of persistent and non-persistent region in the cluster
(between nodes), the above assumption will not hold good; resulting in data
serialization exception.

When there is a mix of persistent and non-persistent region, the version info
is always generated from the persistent member. While constructing the tombstone
message, even though there is no tombstone version generated on non-persistent
member, it was added into the tombstone message, resulting in mixed version
source, causing deserialization failure.