GEODE-2853: Change of locator list request interval
If you connect to a Geode cluster using a locator from the client, the
locator list request will be executed at regular intervals in the
background thread. I want to tune this interval. I understand that this
interval can be changed by the ping-interval of the Pool attribute.
However, I think that ping-interval originally sets the ping interval
for health check to the cache server. Therefore, it is not possible to
change the locator list request interval without changing the health
check interval to the cache server. So,I want to add a java system
property that can change the locator list request interval.

The locator list request interval is determined by the following
priority order.

  1. java system property "gemfire.LOCATOR_UPDATE_INTERVAL"
  2. ping-interval of the Pool attribute

In addition, when changing this time, the background thread is activated
only when this value is a positive value.