GEODE-2875 shutdown is taking as long as 20 seconds
The band-aid fix for this problem was to reduce the wait-time on joining a
thread sending shutdown messages.

This change set alters the membership manager, reviving the path of sending
certain messages like ShutdownMessage over UDP instead of TCP/IP stream
sockets.  This avenue doesn't block trying to form point-to-point
connections so the join() can complete in a short amount of time.