GEODE-2875 shutdown is taking as long as 20 seconds
The fix for this issue causes one of the test cases in LocatorDUnitTest
to fail consistently. With the fix we don't create any TCP/IP connections
in this test during startup but the test expects one to have been created
and it expects the connection's reader-thread to have initiated suspect
processing.

The test needs to be altered to ensure that this thread has been created
by sending message that requires a reply. The reply will be sent over the
expected connection, ensuring that there is a reader-thread.