GEODE-2915 Messages rejected due to unknown "vmkind"
The fix for GEODE_2875 has exacerbated this problem, which we used to only
see in cases where disable-tcp=true or when multicast was enabled.

The problem is that JGroupsMessenger is not sending the "vmkind" of the
sender in message headers.  This part of the header comes from
GMSMember.writeEssentialData().  I've changed it here to include the
vmKind if the recipient isn't using geode 1.0, which doesn't expect
the version byte.