GEODE-2875 shutdown is taking as long as 20 seconds
With a 1.2.0 release pending I am backing out the part of the fix for
this issue that routed certain messages over UDP unicast.  This change
needs more testing as Hitesh suspects it is implicated in a number of
hangs he has seen in his tests.

The Shutdown message is still routed over UDP but all others are now
directed to TCP/IP stream sockets, as they were before.