GEODE-2975: Change the location of the lucene xsd
The lucene xsd was not in the right location on disk, which resulted in
a failure to validate lucene elements in the xml file.

Adding tests and validation for cases where lucene elements show up in the
wrong place, which we were unable to catch with the xsd validation.

This closes #538