GEODE-3027: add a simple PartitionResolver
The new PartitionResolver is:
org.apache.geode.cache.util.StringPrefixPartitionResolver
It has a default no-arg constructor and can be configured
on a partitioned region as its PartitionResolver.
It requires that the region it is on to use String keys
and that the key contains at least one "|" delimiter.
The portion of the key that precedes the delimiter,
called the prefix, will
be returned from getRoutingObject and used to partition
the data on the region.
For example if the key is "customer#36|item#45" then it would
be partitioned uisng "customer#36".
This resolver does not allow the prefix string to contain a
delimiter.